refactor(web): give the position its own feature directory
It was a file inside `lib/app`, which holds the document title, the footer
content and the server-side settings fetch — things about the application
rather than about where in it the user stands. The position is its own
subject, with a type, a class and a hook, so it gets its own directory.

`useAppParams` is deleted rather than moved. Its only caller already held
the search params it wrapped, and called `useSearchParams` on the line
above, so it was a second subscription to the same thing behind a
different name.

Also moves a doc comment back above the function it describes; an earlier
edit inserted a helper between the two.
